Plan: plans/PR-FAQ-Deflection-Report-Rule-File.md
Slice phase: Vertical slice

Adds reusable JSON rule-file support to the customer-facing FAQ deflection report CLI and moves the existing FAQ CLI rule parser into one shared script helper so the two CLIs do not drift.

Intentional:
- JSON rule files only, matching the existing FAQ CLI contract.
- The FAQ CLI is touched to remove duplicate parser ownership while preserving its existing validation tests.
- No docs update in this slice; operator docs can follow after the shared CLI surface lands.

Deferred:
- Future product-polish slice: add report CLI rule-file examples to the extracted README and host runbook.

Parked hardening:
- None.
